metric: refactor types into own files
